requirements.txt
#
# Install requirements
#

click>=3.3,<4
colorama
psutil==2.2.1

bunch==1.0.1
configobj>=5.0,<6
#-e git+https://github.com/jhermann/configobj@move-to-package-32#egg=configobj

#rudiments==0.2.0
-e git+https://github.com/jhermann/rudiments#egg=rudiments

PyYAML
